# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hub Comercial iSolutis — API",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/projetos": {
      "get": {
        "tags": [
          "Projetos"
        ],
        "summary": "Listar",
        "operationId": "listar_api_v1_projetos_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjetoLeitura"
                  },
                  "type": "array",
                  "title": "Response Listar Api V1 Projetos Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Projetos"
        ],
        "summary": "Criar",
        "operationId": "criar_api_v1_projetos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjetoEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjetoLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/projetos/modelo": {
      "get": {
        "tags": [
          "Projetos"
        ],
        "summary": "Modelo",
        "operationId": "modelo_api_v1_projetos_modelo_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "negocio_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Negocio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModeloProjeto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projetos/etapas-padrao": {
      "get": {
        "tags": [
          "Projetos"
        ],
        "summary": "Etapas Padrao",
        "operationId": "etapas_padrao_api_v1_projetos_etapas_padrao_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "inicio",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Inicio"
            }
          },
          {
            "name": "entrega",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entrega"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EtapaSugerida"
                  },
                  "title": "Response Etapas Padrao Api V1 Projetos Etapas Padrao Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projetos/{id_}": {
      "put": {
        "tags": [
          "Projetos"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_api_v1_projetos__id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjetoAtualizar"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjetoLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Projetos"
        ],
        "summary": "Excluir",
        "operationId": "excluir_api_v1_projetos__id___delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projetos/{id_}/relatorio": {
      "get": {
        "tags": [
          "Projetos"
        ],
        "summary": "Relatorio",
        "operationId": "relatorio_api_v1_projetos__id___relatorio_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tarefas": {
      "get": {
        "tags": [
          "Tarefas"
        ],
        "summary": "Listar",
        "operationId": "listar_api_v1_tarefas_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "busca",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Busca"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TarefaLeitura"
                  },
                  "title": "Response Listar Api V1 Tarefas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tarefas"
        ],
        "summary": "Criar",
        "operationId": "criar_api_v1_tarefas_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TarefaEntrada"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TarefaLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tarefas/{id_}": {
      "put": {
        "tags": [
          "Tarefas"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_api_v1_tarefas__id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TarefaAtualizar"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TarefaLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tarefas"
        ],
        "summary": "Excluir",
        "operationId": "excluir_api_v1_tarefas__id___delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tarefas/{id_}/coluna": {
      "patch": {
        "tags": [
          "Tarefas"
        ],
        "summary": "Mover",
        "operationId": "mover_api_v1_tarefas__id___coluna_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoverTarefa"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TarefaLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EtapaSugerida": {
        "properties": {
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "descricao": {
            "type": "string",
            "title": "Descricao"
          },
          "entregaveis": {
            "type": "string",
            "title": "Entregaveis"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "inicio": {
            "type": "string",
            "format": "date",
            "title": "Inicio"
          },
          "fim": {
            "type": "string",
            "format": "date",
            "title": "Fim"
          }
        },
        "type": "object",
        "required": [
          "titulo",
          "descricao",
          "entregaveis",
          "status",
          "inicio",
          "fim"
        ],
        "title": "EtapaSugerida"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModeloProjeto": {
        "properties": {
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "cliente_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cliente Id"
          },
          "negocio_id": {
            "type": "string",
            "format": "uuid",
            "title": "Negocio Id"
          },
          "orcamento_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orcamento Id"
          },
          "responsavel_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsavel Id"
          },
          "inicio": {
            "type": "string",
            "format": "date",
            "title": "Inicio"
          },
          "escopo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escopo"
          },
          "pos_entrega": {
            "type": "string",
            "title": "Pos Entrega"
          }
        },
        "type": "object",
        "required": [
          "titulo",
          "cliente_id",
          "negocio_id",
          "orcamento_id",
          "responsavel_id",
          "inicio",
          "escopo",
          "pos_entrega"
        ],
        "title": "ModeloProjeto",
        "description": "Valores iniciais para um projeto novo aberto a partir de um negócio ganho."
      },
...
      "ProjetoAtualizar": {
        "properties": {
          "titulo": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Titulo"
          },
          "cliente_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cliente Id"
          },
          "negocio_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negocio Id"
          },
          "orcamento_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orcamento Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "planejamento",
              "construcao",
              "validacao",
              "entregue",
              "pausado"
            ],
            "title": "Status",
            "default": "planejamento"
          },
          "responsavel_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsavel Id"
          },
          "inicio": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inicio"
          },
          "entrega": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entrega"
          },
          "objetivo": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Objetivo"
          },
          "escopo": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Escopo"
          },
          "fora_escopo": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Fora Escopo"
          },
          "pos_entrega": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Pos Entrega"
          },
          "etapas": {
            "items": {
              "$ref": "#/components/schemas/EtapaEntrada"
            },
            "type": "array",
            "title": "Etapas",
            "default": []
          },
          "versao": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Versao"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "titulo",
          "cliente_id",
          "versao"
        ],
        "title": "ProjetoAtualizar"
      },
      "ProjetoEntrada": {
        "properties": {
          "titulo": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Titulo"
          },
          "cliente_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cliente Id"
          },
          "negocio_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negocio Id"
          },
          "orcamento_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orcamento Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "planejamento",
              "construcao",
              "validacao",
              "entregue",
              "pausado"
            ],
            "title": "Status",
            "default": "planejamento"
          },
          "responsavel_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsavel Id"
          },
          "inicio": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inicio"
          },
          "entrega": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entrega"
          },
          "objetivo": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Objetivo"
          },
          "escopo": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Escopo"
          },
          "fora_escopo": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Fora Escopo"
          },
          "pos_entrega": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Pos Entrega"
          },
          "etapas": {
            "items": {
              "$ref": "#/components/schemas/EtapaEntrada"
            },
            "type": "array",
            "title": "Etapas",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "titulo",
          "cliente_id"
        ],
        "title": "ProjetoEntrada"
      },
      "ProjetoLeitura": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "atualizado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Atualizado Em"
          },
          "criado_por": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criado Por"
          },
          "atualizado_por": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atualizado Por"
          },
          "versao": {
            "type": "integer",
            "title": "Versao"
          },
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "cliente_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cliente Id"
          },
          "cliente_nome": {
            "type": "string",
            "title": "Cliente Nome"
          },
          "negocio_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negocio Id"
          },
          "orcamento_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orcamento Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "responsavel_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsavel Id"
          },
          "inicio": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inicio"
          },
          "entrega": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entrega"
          },
          "objetivo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objetivo"
          },
          "escopo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escopo"
          },
          "fora_escopo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fora Escopo"
          },
          "pos_entrega": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pos Entrega"
          },
          "progresso": {
            "type": "integer",
            "title": "Progresso"
          },
          "etapas": {
            "items": {
              "$ref": "#/components/schemas/EtapaLeitura"
            },
            "type": "array",
            "title": "Etapas"
          }
        },
        "type": "object",
        "required": [
          "id",
          "criado_em",
          "atualizado_em",
          "criado_por",
          "atualizado_por",
          "versao",
          "titulo",
          "cliente_id",
          "cliente_nome",
          "negocio_id",
          "orcamento_id",
          "status",
          "responsavel_id",
          "inicio",
          "entrega",
          "objetivo",
          "escopo",
          "fora_escopo",
          "pos_entrega",
          "progresso",
          "etapas"
        ],
        "title": "ProjetoLeitura"
      },
...
      "TarefaEntrada": {
        "properties": {
          "titulo": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Titulo"
          },
          "coluna": {
            "type": "string",
            "enum": [
              "a_fazer",
              "fazendo",
              "revisao",
              "concluido"
            ],
            "title": "Coluna",
            "default": "a_fazer"
          },
          "responsavel_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsavel Id"
          },
          "prazo": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prazo"
          },
          "prioridade": {
            "type": "string",
            "enum": [
              "alta",
              "media",
              "baixa"
            ],
            "title": "Prioridade",
            "default": "media"
          },
          "cliente_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cliente Id"
          },
          "projeto_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projeto Id"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "checklist": {
            "items": {
              "$ref": "#/components/schemas/ChecklistEntrada"
            },
            "type": "array",
            "title": "Checklist",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "titulo"
        ],
        "title": "TarefaEntrada"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}